moved errors to own component for sep. of concerns
